# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Music Together - collaborative listening rooms
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CreateRoomBody {
  hostName: string;
}

export interface Room {
  id: string;
  hostName: string;
  createdAt: string;
}

export interface YoutubeVideo {
  videoId: string;
  title: string;
  channelTitle: string;
  thumbnail: string;
  /** @nullable */
  duration: string | null;
}

export type YoutubeSearchParams = {
  q: string;
};