# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Music Together - collaborative listening rooms
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new listening room
 */
export const CreateRoomBody = zod.object({
  hostName: zod.string(),
});

/**
 * @summary Get room info
 */
export const GetRoomParams = zod.object({
  roomId: zod.coerce.string(),
});

export const GetRoomResponse = zod.object({
  id: zod.string(),
  hostName: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Search YouTube videos
 */
export const YoutubeSearchQueryParams = zod.object({
  q: zod.coerce.string(),
});

export const YoutubeSearchResponseItem = zod.object({
  videoId: zod.string(),
  title: zod.string(),
  channelTitle: zod.string(),
  thumbnail: zod.string(),
  duration: zod.string().nullable(),
});
export const YoutubeSearchResponse = zod.array(YoutubeSearchResponseItem);