# lib/api-zod/src/generated/types/youtubeVideo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Music Together - collaborative listening rooms
 * OpenAPI spec version: 0.1.0
 */

export interface YoutubeVideo {
  videoId: string;
  title: string;
  channelTitle: string;
  thumbnail: string;
  /** @nullable */
  duration: string | null;
}